Add flushAllExpected
For some reason, `flush()` no longer flushes subsequent requests that arrive as
a result of the initial flush - presumably because bluebird's interaction with
`setTimeout` is different to q's, somehow.

The whole thing always seemed rather flaky to me. Let's add a
`flushAllExpected` method, which won't return until everything has been
flushed. (We'll probably want to add some more options to this method for
flexibility, but we can do that later.)