Bump default flushAllExpected to 1000ms
Empirically, things are taking longer than 100ms to complete - particularly in
the app-level tests in riot-web.

The only real reason for a timeout here is so that, if the test is killed by
mocha (after 2 seconds by default), we don't end up with a flush going on
forever and spewing logs all over subsequent tests. Hopefully 1s represents a
better compromise between that situation and not killing tests prematurely.

(Ideally we might have a cancelAllPendingFlushes which could be called in
afterEach, but that sounds like work).